+++ teevid-client-native/include/teevid_sdk/IFrameModifier.h
@@ -6,7 +6,7 @@ namespace teevid_sdk
     class IFrameModifier {
      public:
       virtual void OnAudioSourceFrame (unsigned char *data, size_t size, int channels, int bps) = 0;
-      virtual void OnVideoSourceFrame (unsigned char *data, size_t size, size_t stride, int *alpha_data, size_t alpha_size) = 0;
+      virtual void OnVideoSourceFrame (unsigned char *data, size_t size, size_t stride, float *alpha_data, size_t alpha_size) = 0;
     };
 }
 #endif //_IFRAMEMODIFIER_H_
